Tests: Changed test_fpEmptyAttrs expected test result
This test included an empty Attributes token in a footprint. The expected result was, that
the token will create its respective class object in the parsed footprint, and to generate
the same empty (attr  ) token when using 	o_file(). It was changed to expect this
token to be gone after parsing to comply with how KiCad 6.0 handles empty attributes (it
won't include the (attr  ) token at all)