Remove the sanity check on parse_symbol_id.
The sanity check failed for otherwise valid library names, such as "formula:R_0_2512". And the sanity check doesn't seem to provide much protection, as there is already a fallthrough in the "The symbol is a top-level symbol without a library nickname" clause.

Example(s) of where this issue was identified in the wild:
- https://github.com/olin-electric-motorsports/olin-electric-motorsports/blob/main/examples/m.2_example/motherboard/m.2_motherboard.kicad_sch